Renames to generics
