--- src/Ex.Microsoft.Azure.ServiceBus.Hosting/DynamicMessageHandlerOptions.cs
@@ -1,10 +1,0 @@
-﻿using System;
-using System.Threading;
-
-namespace Microsoft.Azure.ServiceBus.Hosting
-{
-    public class DynamicMessageHandlerOptions
-    {
-        public Func<Message, object> MessageDeserializer { get; set; }
-    }
-}
