+++ src/Ex.Microsoft.Azure.ServiceBus.Hosting/Generic/GenericMessageHandlerOptions.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace Microsoft.Azure.ServiceBus.Hosting.Generic
+{
+    public class GenericMessageHandlerOptions
+    {
+        public Func<Message, object> MessageDeserializer { get; set; }
+    }
+}
